Add minor fixes
